feat: Reestructuración de controladores y cambio de tipo de ID a Integer en entidades y repositorios

+++ src/main/java/com/spring/team1/tiendamia/Models/payload/categoria/CategoriaRequestDTO.java
@@ -9,5 +9,5 @@ public class CategoriaRequestDTO {
     private String nombre;
     @NotBlank(message = "El slug de la categoría es obligatorio")
     private String slug;
-    private Long id_categoria_padre = null;
+    private Integer id_categoria_padre = null;
 }
