+++ src/main/java/com/spring/team1/tiendamia/repository/usuario/TokenRecuperacionRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.spring.team1.tiendamia.models.recuperacionPassword.TokenRecuperacion;
 
 @Repository
-public interface TokenRecuperacionRepository extends JpaRepository<TokenRecuperacion, Long> {
+public interface TokenRecuperacionRepository extends JpaRepository<TokenRecuperacion, Integer> {
     Optional<TokenRecuperacion> findByToken(String token);
 }
